requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
graphyte==1.6.0
idna==2.9
numpy==1.18.4
PyYAML==5.3.1
requests==2.23.0
urllib3==1.25.9
